requirements.txt
aiohttp==3.6.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
multidict==4.7.6
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
yarl==1.5.1
